Penulimate Changes
I've got the links fixed, visuals updated, and code cleaned. However, I'm still getting HTML validation errors. One error is with the used of "|" in the google fonts link, and it doesn't go away when I use the HTML code "&vert;". The second is with using bold as a child of a p element, which is the only way I can understand to use it instead of span, as you instructed me to. I've also ommited the use of font shorthand in some places to avoid unecessarily ungrouping specific font stylings, but I can change it if you disagree with that decision.